app converted to run on Postgres database
