+++ db/index.js
@@ -1,51 +1,25 @@
-const mongoose = require('mongoose');
+const { Pool } = require('pg');
 
-const mongoURI = 'mongodb://localhost:27017/similarVehicles';
-
-const db = mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true});
-// mongoose.connect(mongoURI, { useNewUrlParser: true, useUnifiedTopology: true});
-
-db
-  .then(db => console.log(`Connected to: ${mongoURI}`))
-  .catch(err => {
-    console.log(`There was a problem connecting to mongo at: ${mongoURI}`);
-    console.log(err);
-  });
-
-/* The below also works, taken from Mongoose Docs.
- * I prefer the above as it logs more clearly what is happening in the terminal
- */
-// var gooseDB = mongoose.connection;
-
-// gooseDB.on('error', console.error.bind(console, 'Connection error:'));
-// gooseDB.once('Connection Open', () => console.log('Connected to mongoDB'));
-
-const similarVehicleSchema = new mongoose.Schema({
-    year: Number,
-    make: String,
-    model: String,
-    class: String,
-    price: String,
-    miles: String,
-    engine_L_Cyl: String,
-    transmission: String,
-    exterior_color: String,
-    interior_color: String,
-    picture: String
-});
-
-const Vehicle = mongoose.model('Vehicle', similarVehicleSchema);
-
-const addManyVehicles = (cars) => {
-    return new Promise((resolve, reject) => {
-        Vehicle.collection.insertMany(cars, (err, docs) => {
-            if (err) {
-                reject(err)
-            } else {
-                resolve(docs.length, 'added');
-            }
-        })
-    })
+const config = {
+    database: 'similar_vehicles',
+    user: process.env.DB_USER,
+    host: process.env.DB_HOST,
+    port: process.env.DB_PORT,
+    password: process.env.DB_PASS
 };
 
-module.exports = addManyVehicles;
+const pool = new Pool(config);
+
+// pool.query('select * from similar_vehicles where id < 5 LIMIT 1', (err, res) => {
+//     if (err) console.log('ERROR: ', err);
+
+//     pool.end()
+// })
+
+
+module.exports = {
+    query: (text, params, callback) => {
+        // console.log(text);
+      return pool.query(text, params, callback)
+    },
+};
